requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
Werkzeug==3.0.3
python-dotenv==1.0.1
passlib==1.7.4
PyJWT==2.9.0
pytest==8.3.2
pytest-cov==5.0.0
coverage==7.6.4
line_profiler==4.1.3
memory-profiler==0.61.0
psutil==5.9.8
sphinx==7.3.7
sphinx_rtd_theme==2.0.0
requests==2.32.3
pybreaker==1.0.0
